my_webapp/requirements.txt
aiogram>=3.0
fastapi
uvicorn
sqlalchemy
aiosqlite
jinja2
python-dotenv
pydantic
